app/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
diff-match-patch==20230430
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-filter==2.4.0
django-import-export==4.0.8
django-redis==5.4.0
django-simple-history==3.6.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
hiredis==2.3.2
httpagentparser==1.9.5
kombu==5.3.7
prompt_toolkit==3.0.47
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.6
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
pylint-django==2.5.5
pylint==3.2.4
black==24.4.2